requirements.txt
django==5.0.6
gunicorn==22.0.0
psycopg2-binary==2.9.9
djangorestframework==3.15.2
djangorestframework_simplejwt==5.3.1
django-ipware==7.0.1